Added hotplugging protection to TRRS
Technically added hotplugging protection. But it won't work because
current isn't able to flow both directions in this design. It would
require one side to always be the slave. But the hotplug protection for
the master would then suck. So I'm about to try switch to USB-C for the
interconnect.